ATerm cleanup
* Remove some unused code
* Remove references to ATerror in mCRL2 specific code
* Add some (ugly) explicit casts in aterm code.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@9087 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ libraries/aterm/include/aterm/aterm2.h
@@ -155,7 +155,7 @@ namespace aterm
   inline
   int ATgetInt(const ATermInt t)
   {
-    return t->aterm.value;
+    return (int)t->aterm.value; // See comment in definition __ATermInt
   }
 
   /* The ATermAppl type */
